web-client/requirements.txt
frozendict
requests
inflection
urllib3
python-dateutil
pydantic
fastapi
ujson
jinja2
aiofiles
pytz
json2html
pandas
python-slugify
python-multipart
async-cache
async_to_sync
git+https://github.com/nadirizr/json-logic-py.git@master
git+https://github.com/archetipo/python-formio-data.git@component_for_jinja2


